a bit of cleaning up
